Move code only used for Wine detection into wine file
